Load md file for each component
