+++ app/src/components/ProtvistaSequence.jsx
@@ -1,6 +1,8 @@
 import React, { Fragment, Component } from "react";
 import ProtvistaSequence from "protvista-sequence";
 import loadWebComponent from "../utils/load-web-component";
+import Readme from './Readme';
+import readmeContent from '../../../packages/protvista-sequence/README.md'
 
 class ProtvistaSequenceWrapper extends Component {
   componentDidMount() {
@@ -16,6 +18,7 @@ class ProtvistaSequenceWrapper extends Component {
     loadWebComponent("protvista-sequence", ProtvistaSequence);
     return (
       <Fragment>
+        <Readme content={readmeContent}/>
         <protvista-sequence
           id="seq1"
           length="223"
