+++ app/src/components/ProtvistaTrack.jsx
@@ -3,6 +3,8 @@ import DataLoader from "data-loader";
 import ProtvistaFeatureAdapter from "protvista-feature-adapter";
 import ProtvistaTrack from "protvista-track";
 import loadWebComponent from "../utils/load-web-component";
+import Readme from './Readme';
+import readmeContent from '../../../packages/protvista-track/README.md'
 
 import data from "../mocks/features.json";
 import data2 from "../mocks/features2.json";
@@ -23,6 +25,7 @@ class ProtvistaTrackWrapper extends Component {
     loadWebComponent("data-loader", DataLoader);
     return (
       <Fragment>
+        <Readme content={readmeContent} />
         <h2>Track with data-loader</h2>
         <protvista-track
           length="770"
